requirements.txt
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
globus-sdk==3.28.0
idna==3.4
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
pycparser==2.21
PyJWT==2.8.0
python-json-logger==2.0.7
PyYAML==5.4.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.2
structlog==23.1.0
tabulate==0.9.0
urllib3==2.0.4
watchdog==3.0.0
